Attempt to fix issues with Chrome
Instead of the button triggering the image being created and downloaded,
the image is created by the Draw Page button, with the link updated at
that point. The link points to the data: URL, and with the download
attribute provides an acceptable name for the provided file.

A button is embedded inside the link for appearance.